requirements.txt
# 核心依赖
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.12
starlette==0.46.2
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1

# HTTP 相关
h11==0.16.0
httpcore==1.0.9
httpx==0.27.1
httptools==0.6.4
httpx-sse==0.4.0
requests==2.32.3
urllib3==2.4.0

# MCP 相关
mcp[cli]==1.7.1
sse-starlette==2.3.4

# 工具依赖
python-dotenv==1.1.0
PyYAML==6.0.2
typing_extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5

# Supabase 相关
supabase==2.4.3